requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.36
openai==1.6.1
httpx==0.25.2
python-multipart==0.0.6
twilio==8.10.0
requests==2.31.0
elevenlabs
psycopg2-binary
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]